Update README.md
User ID's for solarcoin affiliate websites. And can be used in conjunction with Blockpass.org. user ID system. 
Open source JSON format. This reason for adding user ID is to protect the privacy of having all information open. User ID is only known to participating affiliate website. In the future will be stored on a decentralized cloud.